account for gaps between sectiotns
due to requirements in the alignment of sections there may be 'gaps' between
sections. That is, given two contiguous sections A and B it may be that `addr(A)
+ sizeof(A) != addr(B)` but rather `addr(A) + sizeof(A) == addr(B) + 2` if for
example `alignof(A) == 2` and `alignof(B) == 4`. Note that these gaps may still
exist even if `. = ALIGN(4)` is used to align the boundaries of a section in the
linker script: the section may have a bigger alignment (due to its contents)
than the one that was specified in the linker script

this commit also switches from xmas-elf to goblin; the former does not expose
the alignment of sections (`sh_addralign`)